RTE Model refactoring: (#494)
* RTE Model refactoring:

derive RteItem from XmlTreeItem<RteItem>

make VISIT result class enum

remove duplicated methods from RteItem

use size_t instead of int for collection sizes

add RteComponentInstance::GetComponent() override to simplify logic

simplify XmlTreeItem::AddChild()

+++ libs/rtemodel/include/RteDevice.h
@@ -1371,14 +1371,14 @@ public:
    * @brief get number of RteDeviceItemAggregate children
    * @return number of RteDeviceItemAggregate children as integer
   */
-  int GetChildCount() const { return (int)m_children.size(); }
+  size_t GetChildCount() const { return m_children.size(); }
 
   /**
    * @brief get number of RteDeviceItemAggregate children of requested hierarchy type
    * @param type RteDeviceItem::TYPE
    * @return  number of RteDeviceItemAggregate children as integer
   */
-  int GetChildCount(RteDeviceItem::TYPE type) const;
+  size_t GetChildCount(RteDeviceItem::TYPE type) const;
   /**
    * @brief create short device item description
    * @return summary string
